Update patch set 1
Patch Set 1:

Pam, to your question, not sure if there is any redundancy of projects between drools-application and this patch in policy-engine repo from Ravi.   Ideally, should avoid duplication.  From drools pdp side, any dependent jars that the "rules" need for proper operation must be in nexus repo.    If they are in drools-application, they will be installed automatically in nexus, as part of the drools installation.

Patch-set: 1
Label: Verified=0