Update patch set 1
Patch Set 1:

Pamela, I created a Jira ticket for this user story today.  POLICY-67 can be viewed at this link...
https://jira.onap.org/browse/POLICY-67?filter=-1

The description provides some information with wiki links provided for the feature level and the user stories that I created to provide a solution for the feature.  It is important to note that this is the first part of the feature, I will be working on enhancing the getDecision API in the current iteration to provide the desired treatment in the details of a "PERMIT" response. 

Please let me know when will be a good time for us to meet and I can provide an overview and demo of the feature.

Patch-set: 1
Label: Code-Review=0