Update patch set 1
Patch Set 1:

Error seems related to an h2 database lock timeout condition.   There has some outage on LF servers, so wondering that's real root cause, better scenario to deal with that if it turned out to be a timing issue.    We'll see.  Kicking the verify job again, now that there is reports that the power outage issues has been mitigated (not totally).

Patch-set: 1