Update patch set 1
Patch Set 1:

It is not clear if it should be 1.3.0-SNAPSHOT (which would be very unusual to be the only project in onap) vs. 1.1.0 official the claim, or simply just the fact that 1.1.0-SNAPSHOT is missing from nessus.   I linked this ticket to PORTAL-33.  Merging to fix build problems, but may need to update version.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0