Update patch set 1
Patch Set 1:

Pam, given the review is already out, I don't see a good solution for this, other than take it for review.   I would assume to split this commit in multiple will be quite a large task for Tarun (is this correct assumption Tarun).   The comment here Tarun, is that the submit is quite large and in unrelated areas, from rest apis, to brms gw, to elasticsearch, .. which makes it difficult to review, and as Pam mentioned does not follow ONAP guidelines.   Provided the split work of the code would be painful, and some fixes are necessary for the load.  I think we should accept it for review this time.  Pam, will you be okay with that or think it is calling for project management trouble down the line?

Patch-set: 1