Update patch set 2
Patch Set 2:

Ravi, can you amend it to say POLICY-117 in upper case in title.  I noticed that using the upper case convention allows to link both gerrit and JIRA with links, otherwise it doesn't.   Bottomline, JIRA tickets are case sensitive, it has to match the upper case convention to be able to navigate between the 2 systems.

Patch-set: 2
Label: Code-Review=0