Update patch set 2
Patch Set 2:

Pam, regarding the ECOMP-ControlloopPolicy we may need to test it before to make sure our changes to remove and replace. There are some changes done in ECOMP-ControlloopPolicy which can affect the policy PAP and CLAMP using it. 
So can we allow this commit meanwhile, we can test if we can use the new ControlloopPolicy in policy/drools-application and make that changes in a new commit.

Patch-set: 2
Label: Code-Review=0