Update patch set 3
Patch Set 3:

Oh, I think the issue is that "TestUtils" is in the "test" directory and the contents of the test directory does not get published with the jar with maven install.  I think there is something that can be done about it in the pom to force them to include but in this case, seems reasonable providing that these are junits to copy parameters in this file.

Patch-set: 3
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0