Update patch set 2
Patch Set 2:

Pam, yes this is necessary.  The API support is a part of the platform for all policy types.  We don't want a platform that has API support for some policy types and not for others.  This is a new feature and the chance of it conflicting with anything is close to 0%.

Patch-set: 2
Reviewer: Gerrit User 805 <805@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0