Update patch set 1
Patch Set 1:

The other changes in the *.properties and *.json files are not necessary and should be working same as in R0 where we have these scripts converting to the right version from $POLICY_HOME/etc/build.info.   

See:  https://git.onap.org/policy/docker/tree/config/pe/brmsgw-tweaks.sh

Now, the problem is that the $POLICY_HOME/etc/build.info is not generated anymore in policy/engine build so it is not available at runtime.   This is a bug that has been introduced in R1, and was working OK in R0.

The culprit seems to be related to these changes.   Note the unix_pe_version profile in R0:

https://git.onap.org/policy/engine/tree/packages/base/pom.xml?h=release-1.0.0

compare it with the 1.1.0 version which has been deleted most of it.

https://git.onap.org/policy/engine/tree/packages/base/pom.xml

Patch-set: 1