Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Mike - this is a significant code change. I'd like to ensure that we are not introducing more technical debt and want to see a before and after sonar report on this first. for example, I see a lot of unnecessary variable initializations just off the top of my eye.

Patch-set: 1
Label: Code-Review=-1