Update patch set 1
Patch Set 1:

Mike, we would need the code changes in amsterdam release as well.   

So this change needs to be cherry picked into amsterdam release once merged.

For any submissions in the amsterdam release testing should be very extensive.

Can you confirm specifically that the following test works correctly with these changes?

1. create policy X version i
2. push policy X version i
3. modify policy X (to version i+1)
4. push policy X (version i+1)

As a reference, see the PDP API operations in this script.   Previous to running this script in an ONAP configuration there would be an existing policy that this script modifies and distributes.

https://git.onap.org/demo/tree/vnfs/vFW/scripts/update-vfw-op-policy.sh

That has to be working well in Amsterdam release.

Can you confirm that this test works ok?

Then port it to Amsterdam after we merge it, and verify that the PDP API operations work properly when there are multiple versions.

Patch-set: 1