Update patch set 1
Patch Set 1:

ElkConnectorImpl.java, ElkConnectorImplTest.java, and elasticsearch.yml are part of code that I added to disable ELK dynamic scripting and check that user queries are scrubbed. This was a CSO item. However, I don't understand the changes in the other files. I know they are some kind of enhancement but I'm not sure exactly what behavior it changes.

Patch-set: 1