Update patch set 4
Patch Set 4: Code-Review-1

There is really no need to change this... its example code for clients to call the Java API and create policies.  Its not functional code so changing it will do nothing.  We give this to clients as a starting point to help setup their own client or test.  Here if isEdit is false then we call createPolicy() and if isEdit is true then we call updatePolicy().

Patch-set: 4