Update patch set 4
Patch Set 4:

If we modify as suggested like below
 static Boolean isEdit = new Random().nextBoolean();
Then if initially only if the Random function returns true then update function will be executed then it will be problem because first create has to be called.

So my modification is when create function executes immediately am making the value to true, so next time the update function will be executed. pls check once.

Patch-set: 4