Update patch set 1
Patch Set 1:

Please also be aware there is a new project in ONAP for "Certificate and Secret" management. Sounds very cia-like but it might be prudent to consult that team as well as OOM on how we should be formalizing this work. I hate to have to undo it in a month.

Patch-set: 1