Update patch set 1
Patch Set 1:

I think the base changes of having a certificate store in pap/pdp is needed.   That's necessary for the jvm to operate with certificates regardless for any https communication between components.   

Not sure if we can do that as bug since it is basic that Temoc could complete, or make it part of security user stories for beijing, which I agree we have to figure out.  

Do you have a preference, Pam?

Patch-set: 1