remove mariadb superuser sql statements
PE installation as part of data upgrade invokes
"set global" statements that are typically used
by priviledged accounts as the "root" mariadb user.

As part of the effort to clean up the mariadb docker install,
only permissions to a regular mariadb user (policy_user by default)
are given in 3 databases: onap_sdk, log, and support.
This account cannot do changes outside those.

In any case the usage of the global timeouts is questionable.
We should keep an eye if there's any potential bad side effect,
and correct it if it doesn't.   It's preferable to keep
"policy_user" with a restricted set of privileges, instead
of being a superuser (otherwise can plainly user the root account).

Change-Id: I17b1a6a2cb62a319464e1db4b3aafbc1dbf20d78
Issue-ID: POLICY-650
Signed-off-by: Jorge Hernandez <jh1730@att.com>