Update patch set 1
Patch Set 1:

Michael - this is great. I was wondering if we needed the prepare-package anymore. I actually don't think we do. We are going to switch the CI/CD jobs (upcoming commit). thus:
mvn clean deploy 
By default the ci-management jobs for Policy look for the scripts in the root directory, but I can easily change that.

Patch-set: 1
Label: Verified=0