Update patch set 1
Patch Set 1:

We need to remember before releasing to move all these third party dependencies here and in pdp-d's related repos to released ones.   We should open a JIRA ticket for it.   The maven version plugin, I believe does not replace these third party ones with the corresponding release, so it is a manual operation.

Patch-set: 1