Update patch set 1
Patch Set 1:

I meant replace the:

"FROM ... policy-base" (first line of Dockerfile) with the actual contents, and discontinue policy-base in the docker repo as a possibility.  I think the Dockerfile from policy-base is minimum.  I question its usefulness.

There may some other options as setting the docker registry explicitly to nexus ..

We can submit this, but I think should be revisited in the future, not comfortable with in-flight edits on a Dockerfile.

Patch-set: 1