Update patch set 3
Patch Set 3:

> Question Praveen, the request was to not add anything to the drl,
 > there cannot be any assumptions on its drl content.   Is this what
 > this review is trying to accomplish?

This removes the auto rule generation by pap. We still need the "declare Params" in template to drive the UI to prompt for inputs. For subsequent edit/view from UI, the inputs are stored in the comments of the drl. The drl comment which earlier was used to store templateName only, is now being used for input params. For example, /* Autogenerated Code Please Don't change/remove this comment section. This is for the UI purpose.  	 <$%BRMSParamTemplate=test%$>  <%$Values=controlLoopYaml:-:test:|:operationTopic:-:test:|:notificationTopic:-:test:|:closedLoopControlName:-:test:|:$%>  */

Patch-set: 3
Label: Code-Review=0