Update patch set 1
Patch Set 1:

OK. I'll do that and submit a new review in a minute.

I'm sure PAPServices data clearing is needed and is unrelated to the NotificationService thing.

Actually its funny because the other tests in that class also call the notification service but they don't seem to trigger loading of that class. Pam maybe open a Jira to have a look at this later.

Patch-set: 1