Add database initialization to XACML PDP test
This issue arose due to the order of execution of JUnit tests.

The tests in XACMLPAPTest failed becasue the database was not initialized.
Up till now, the @Before method in the test class
org.onap.policy.pap.ia.DbAuditCompareEntriesTesti initiated the database.
THis bug fix adds a similar @Before method to the XACMLPAPTest class
to initialize the database.

All other classes in the module were tested with both @Before methods
commented out to ensure that no other tests will fail due to this issue.

Issue-ID: POLICY-707
Change-Id: Ibce5fad5c006f7a92b3ef77bde4dc5ed6f7039ed
Signed-off-by: liamfallon <liam.fallon@ericsson.com>