Update patch set 1
Patch Set 1: Code-Review+1

I saw this error in the console log in regards to the tomcat version in the failed csit build as well but wasn't sure if it was relevant.  But I still think there may be some other database init issues as well.  I hope this fixes it.

Patch-set: 1
Label: Code-Review=+1