Update patch set 3
Patch Set 3:

Frank, Ravi, can somebody clarify what this is supposed to do, and against what type of data, is it a maven group id.   It pretty much accepts any arbitrary type of text including spaces, and of any arbitrary length?   If it is a check for something, it is very weak.

Patch-set: 3