OOF Policy Config File Creation Issue
Fixed the issue with creating the config file with the wrong file name
so that the PDP can pick it up when pushing the policy and getCOnfig
will return the Config retrieved rather than a CONFIG_NOT_FOUND 400
status.

I found this issue in the new CSIT tests for HPA policies and this
should also resolve the failed test in Integration as well.

Change-Id: I617133a188f3ba915feafd79da4f3e9d0b0eb4a6
Issue-ID: POLICY-738
Signed-off-by: Michael Mokry <mm117s@att.com>