Update patch set 4
Patch Set 4:

I was actually not sure we had to do the some of the CSS from 3rd party libs like fusion, bower and bootstrap. Should have mentioned that to you yesterday in my rush. I wanted to ask LF if that was necessary. Is it easy to pull those out until we get an answer to the question?

Patch-set: 4
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0