Implement Sonar Nexus REST interface directly
The BRMS gateway used the obsolete nexus-rest-client-java library
to query the Nexus repository. This change removes that dependency and
implements an interface directly to Nexus servers using the Nexus
REST interface.

This change is being submitted to ensure that the security issues with the
plugin are resolved. The unit test for the new code will be implemented
in a later submission.

Issue-ID: POLICY-700
Change-Id: Ia1044e432d370034fe13f42cf7ed55d75e21187a
Signed-off-by: liamfallon <liam.fallon@ericsson.com>