Add JUnit for direct Sonar Nexus REST interface
JUnit tests and changes to the source code to fix issues that were
identified in JUnit testing.

JUnit testing runs tests towards the Nexus server at
https://nexus.onap.org
If that server does not exist then the unit tests will fail
but as the nexus serveir is required for the build to work then
it should be OK to use that server for unit testing. This removes
the need to write complex (and less effective) Mockito type
unit tests.

Issue-ID: POLICY-700
Change-Id: I4aa33a1022033b42bb1fd1fa36785a8e0216b887
Signed-off-by: liamfallon <liam.fallon@ericsson.com>