Update patch set 2
Patch Set 2:

When I changed these classes I was worried that it wasn't thread safe. The old class was pretty much a static class but the new library insists on an instance existing because it's implemented by specializing an abstract class.

The current implementation is a singleton but I'm not sure that the singleton doesn't get overwritten on new connections. If it does, we'd get all sorts of interesting side effects. I don't really understand how the class is being called in the engine.

If its OK with you I'd like to have a look at it in a little more detail later tonight/in the morning.

Patch-set: 2
Label: Verified=0