Update patch set 1
Patch Set 1:

Liam Code Review Comments:
=========================
(6 comments)

I checked the two classes this morning. I think they are fine except for the minor issue with closing the old connection when the URL changes.
PolicyEngineAPI/src/main/java/org/onap/policy/std/AutoClientEnd.java
Line 27:

We don't need this import for the new WS library
Line 38:

We don't need this annotation for the new WS library
Line 139:

OK, start does nothing if the client is already started :-)
Line 184:

I think we need to do a stop() before changing to the new URL, otherwise when we call start() we'll overwrite the URL on line  136 and then we'll come to the if statement on line 138/139 and see that "AutoClientEnd.client != null" and finish. We'll have the situation with the client configured with the new URL but still happily connected to the web socket at the old URL.
PolicyEngineAPI/src/main/java/org/onap/policy/std/ManualClientEnd.java
Line 25:

We don't need this import for the new WS library
Line 36:

We don't need this annotation for the new WS library

Patch-set: 1
Reviewer: Gerrit User 2757 <2757@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0