Update patch set 2
Patch Set 2:

I got this one on my laptop when doing the other fix on the BRMS Gateway unit tests. I thought it was a local thing on my machine becasue when I ran a draft build on Gerrit it worked. It must be an intermittent thing or maybe a unit test order dependency.

Patch-set: 2