Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Remove POLICY-758 from first commit line (per onap guidelines should not contain the ticket number, it is in the Issue-ID).   First line should be a short descriptive message on what the review is doing.

Main issue is with the logback.xml, which is a drastic replacement of a file.   There are some changes that are OK, but many that are not.   Among the ones that are not is to replace ONAP with ECOMP, and the appenders configuration, do not match ONAP guidelines.

Patch-set: 1
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1