fix PE DB lock timeout problems
see companion docker review.

The failOverReadOnly .. flags should only be used when a DB high
availability environment is used, such as a master/slave
configuration.

As they are the jdbc parameters have caused bad interactions
with the underlying persistence libraries rendering the PAP
useless.

It is thought that this may perhaps solve
POLICY-773, POLICY-775, POLICY-779, and perhaps POLICY-776.

Change-Id: Icb69a3afc0a19969a34af40d960fa8e1b94aabc0
Issue-ID: POLICY-777
Signed-off-by: Jorge Hernandez <jh1730@att.com>