Configuration as argument to BRMS Gateway
Added the ability specify the parameter file for the BRMS Gateway
as a argument to allow different configurations to be used during
unit test. This will allow more thorought unit tests to be written.

Replaced System.exit() calls with exeception throws becasue
System.exit() call bring down the entire JVM during testing, terminating
the test at that point.

Changed the package path on four unit test files to the correct path for
unit tests for the BRMS gateway.

Added a unit test for sanity test of the configuraiton file argument.

Issue-ID: POLICY-773
Change-Id: Ic095a5131ddb846eaf3b11157853fab71908c629
Signed-off-by: liamfallon <liam.fallon@ericsson.com>