Upgrade SNAPSHOT to 1.2.2
These reviews must be merged in order because of layering.

This review should be merged SECOND

Note: In order to get past the following layering problems, there
are temporary chages to the following POMs to temporarily use 1.2.1-SNAPSHOT
for those dependencies. They are
ONAP-PAP-REST/pom.xml
 org.onap.policy.drools-applications.controlloop.common.policy-yaml
ONAP-PDP/pom.xml
 org.onap.policy.drools-pdp.policy-endpoints

A separate review will be used to restore these dependencies to 1.2.2-SNAPSHOT
once the drools-pdp and drools-applications reviews are submitted and merged

PLD added version.properties

Issue-ID: POLICY-798
Change-Id: Id7e83f2ba12181c63cc8845b7eeccf2004d2bab9
Signed-off-by: liamfallon <liam.fallon@ericsson.com>
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>
Signed-off-by: liamfallon <liam.fallon@ericsson.com>