Update patch set 1
Patch Set 1: Code-Review-2

This is 10,000 lines of code. It will take a long time to review this and it needs to be presented to the ONAP community in order to avoid "upstream first" accusations.

Praveen, please be available to present to the community this feature on 7/11/2018.

I am doing -2 to avoid accidentally merging this. It would be better to have this as a draft as we go through this.

Patch-set: 1
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2