Update patch set 1
Patch Set 1:

(2 comments)

Yes, it's hard to see the changed try-with-resources in all the checkstyle changes, I think I spotted them all but am not sure. If Krishnajinka could just give us the line numbers where the try-with-resources are in each class it would help with this review.

Patch-set: 1
Label: Verified=0