Update patch set 6
Patch Set 6:

(5 comments)

Thanks very much for the work on this change, we just need to fix the exception passing to the loggers pass the review.

I also noticed that there is excellent Javadoc on most of the methods but its missing on a few. If you have time to add javadoc to those it would be great. The Eclipse plugin JAutodoc is very good at generating Javadoc.

Patch-set: 6
Label: Code-Review=0