Update patch set 2
Patch Set 2: Code-Review+1

So I think that the change you made to making a variable with a meaningful name is enough for this code review. Assuming Liam is ok with it.

Perhaps we log a JIRA on this to go back later and build in property file support.

Patch-set: 2
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1