Update patch set 3
Patch Set 3:

I'd like to get this pushed today so the Integration team and security sub committee can move forward. Please ensure that the versions that are in policy/parent are ok here. There were some upgrades that occurred vs removing duplicated versions. If this fails, we can override versions in a subsequent review if we don't identify any today.

Patch-set: 3
Label: Code-Review=0