Update patch set 8
Patch Set 8:

Ravi, can this work be described either in ticket or in wiki page, having the ticket pointing to the wiki page?   We will need to document it at the end of the release.   Key things to explain are what are raw policies and what they provide on top of the existing policy types.

Patch-set: 8
Label: Code-Review=0