Update patch set 1
Patch Set 1:

The guard policy can be written many ways - its an art. We can revisit it at a later date. Giving the combining algorithm, we probably only need one rule to detect permit or deny. Whatever is easiest to write, which in the case of XACML isn't always that clear anyway.

Patch-set: 1
Label: Verified=0