Update patch set 2
Patch Set 2:

Yes this is a pain - they claim that they have initiated a release for this with the LF. But its not there yet. I don't want to compile to staging, because ultimately those artifacts disappear. And then we are stuck. Let's hold this for the moment. I will ping them.

Patch-set: 2