Remove override of parent aaf version
policy/parent now declares the aaf dependency so consistency
is maintained in all the repositories. Removing the version
declaration in order to pull in the declared version.

The constructor changes to add a namespace. A null value is
still possible to be used. If we are given a namespace then we
will have to add that in. Preferrably via a properties file.

I also noticed there is a hard-coded URL in here in the
setupAAF() method on line 212. That should be moved to a property
file.

Issue-ID: POLICY-1119
Change-Id: I97dcd7d669d3d05db7c6597be6f3038183fa4e17
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>