CADI AAF changes for policy/engine
Added cadi properties and modified policy aaf client code.

Fixed issue with namespace and modified code to reverse it for
structuring the username sent in AAF API call

Added properties for keystore and keystore password after getting error
response from AAF when setting up the connection to AAF

Missed a fix for one of Jorge's comments in last patch, here it is.

Change-Id: Ic164ade8aa34da95a560c1592656e0caf990a595
Issue-ID: POLICY-913
Signed-off-by: Michael Mokry <mm117s@att.com>