Update patch set 1
Patch Set 1:

I see Mike, existing seed code, has values wrong.   But nevertheless, they are totally off, and they should be corrected.   The question is doing it now or as a high priority bug ticket in the very near future or now.   If it takes you little time I would prefer that you corrected now, otherwise we can discuss.   But I'll leave it open to you and other to pitch in, and convey their opinion.

In general, the namespace is org.onap.policy, the "policy.onap.org" is more like the "dns" part that does under the user id policy (it's not a namespace, but rather a namespace locator for the user id).    The code does not need a property for that (policy.aaf.resource) neither as they can be reversed from the "real" namespace with a simple function, and therefore require less configuration (policy.onap.org is just the reverse of the namespace).

I think the "policy." prefix, that is ok, thinking more about it, as I noted that can be used to distinguish "policy" component properties from "aaf" component properties that get passed to AAF library.  So, you can leave as is, and also use it to prefix ""aaf.root.permission" for consistency.

Let me know your thoughts.

Patch-set: 1