Update patch set 1
Patch Set 1: Code-Review-1

Hi Abhay.  By design, we only created a hyperlink for JIRA bugs and not for the epics or stories.  (We do have a link which displays all the Policy Beijing Jiras).  Do you know why this is changing now?  
I don't think we should change release notes related to prior releases.  If you think it is necessary to hyperlink the epics and stories, let me know and we can do this beginning with the Casablanca release.
Also, I glanced thru your changes and it looks like you may have added an incorrect for POLICY-63 which happens to be a Jira delivered in Amsterdam.
In my opinion, this Jira should be marked as "Not a Bug" and no changes should be made.

Patch-set: 1
Label: Code-Review=-1