Update patch set 1
Patch Set 1: Code-Review-1

Abhay, The hyperlink for POLICY-63 does not look correct (this was delivered in the Amsterdam Release).  Also, in my opinion, we should not modify the information for prior releases.  If you feel it is necessary to individually hyperlink the epics and stories, let me know and we can look at doing that for the Casablanca release.

Patch-set: 1
Label: Code-Review=-1