Update patch set 3
Patch Set 3: Code-Review+1

This change should fix the problem, if it's due to port-already-in-use.  It should also fix the problem, if the listener hasn't opened the port before the client tried to connect.  If it's neither of these, then it should now at least log an error message to the console to help in identifying the problem.

Patch-set: 3
Label: Code-Review=+1