Update patch set 1
Patch Set 1:

This line has not relevance in ONAP installations, the configuration is taken from base.conf defined in policy/docker (heat) and oom (k8s) installs.   In ONAP, both heat and kubernetes, only the open-jdk is available (not the oracle one).

This line would only come into play when somebody does a "test" installation on his/her PC (with no containers).   In any case, in that sort of installation, the person that is installing is supposed to configure this file (base.conf) per what their system looks like and then proceed with the installation.

I believe the variable should be there, so if somebody does a non-container test installation they remember to change it to whatever location of the jdk they want to use.   We all agree that is better to set it up to the open jdk, even though it is just for reference.   In any case, somebody working on a mac and doing their own testing this way, would be expected to change it to fit their own testing environment.

Patch-set: 1