Update patch set 1
Patch Set 1: -Code-Review

Ok - can we put a comment in with this information so people know why its there? People do download the source and compile locally so they would have to know this. So Jim, is this the reason the MariaDB is not coming up? That's the JIRA that is being used for this submission.

Patch-set: 1
Reviewer: Gerrit User 57 <57@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0