Update patch set 5
Patch Set 5:

I disagree in a separate jira for csit only, if the CSIT is created to support this API, then we use the same JIRA for that submission to show all the work for bringing this API in. It can be a sub-task of this user story if you really want something different.

Patch-set: 5