Remove javascript from sonar
Per Catherine, Dublin will not require javascript for sonar
coverage statistics. We'll add this back in for El Alto if
we determine this repo still needs development.

If LF turns off javascript in the Sonar application then I'll
change this back. But I need to determine if policy/engine has
55% coverage for tomorrow M2.

Issue-ID: POLICY-1536
Change-Id: Ie091e231e1150da3217e10eb28fdbcd7e2ed8320
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>