Update patch set 2
Patch Set 2:

Thanks Praveen, can you add a commented out line to the properties file you mentioned so folks know how to do it.   We'll also need some documentation on how to use the capability, for example how to generate the AES Key and what needs to be done for an admin to go and use it.   This ideally should go in readthedocs, or perhaps in wiki page.   May be Pam has a suggestion on where to put it.   For this review, if you can just add those comments in those properties files.

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0