Update patch set 4
Patch Set 4:

Hello Utkarsh, can you open a JIRA as Jim suggest and feed the number here? We can add it to the backlog, and we can merge.   First one question though to make sure with regards to the xacml.pap.properties - the official copy is under packages, but only this dev. copy has been modified.   Do the same changes that you have in the xacml.pap.properties here need to be applied to the official one?
   
We need to make sure that everything works out of the box in ONAP, and shouldn't have useless parameters neither.

Patch-set: 4