Update patch set 1
Patch Set 1: Code-Review-1

For *new code*, comments should be addressed before merged, either discussed and reach an agreement with with the reviewer or address them in code.   For *existing code*, the comment is not binding, it does not need to be addressed at the time of review, but the contributor should consider it for future work, perhaps track it under a jira ticket.   There should also be junit code coverage for new functionality added.  The inspection is lacking tests.

Patch-set: 1
Label: Code-Review=-1