Update patch set 6
Patch Set 6:

As a general comment, it is better not to duplicate testing configuration all over.   I think this is the only repository that does that and commit those files under source control such as the policy-*store, *cadi keys, etc .. they are already in the packages/ dir.   One solution is to add a .gitignore to ignore those files, even you may use them for local testing.

Patch-set: 6