Update patch set 2
Patch Set 2:

Do you know this issue, when trying to look at logs in the container,  the default vi editor, tail, etc .. don't wrap lines - LINES and COLUMNS are set up correctly?   Do you know how to get pass that, as this is important for debugging at this stage.

Patch-set: 2