Update patch set 2
Patch Set 2:

Thanks for the info about vim, I suspected that was the case, but thought the package would be smaller, we can table it for now.

I am not sure about the "apk upgrade" if it upgrades the OS packages, and you fix the image build against the 1.4.0 that in theory should be blessed, it pretty much defeats the purpose of having a blessed release image to avoid the recent problems we had with openjdk mutability.  I question is if a recommended practice too.

I think all the OS stuff should go to blessed parent images though to fix them in the same version and avoid mutability.   Only application stuff should be left here.  But we don't need to do that right now, may be we can walk towards that in the future.

nss removal can wait too.

Patch-set: 2
Label: Code-Review=0