Update patch set 3
Patch Set 3: Code-Review-1

Need a resolution on the "upgrade" approach if it can be investigated best practice when we are inheriting from a fixed blessed image before letting this one proceed. Intuitively, to me doesn't seem right but I'd like to hear feedback and best practices on this.   I usually haven't come across upgrade operations in packaging in Dockerfiles.

Patch-set: 3
Label: Code-Review=-1