Update patch set 2
Patch Set 2:

Some background team on Mike's work.   These are tentative changes so to Mike can do testing, there does not seem not to be an effective way to test these libraries outside an actual OOM installation.   The changes will require additional work as what Pam is recommending if Mike confirms that it works, or they may be pulled out all together depending on outcome of Mike's tests.   FYI, just some background on side conversations we had, and origin of review.

Patch-set: 2