Update patch set 2
Patch Set 2: Code-Review+2

More thinking about this repo .. I get confused easily with this one.   I think provided that the changes are in the configuration files which are cooked in the image (and not actual Dockerfile or pom.xml dependencies). I think we could go ahead and release it right away, I mean the released artifacts and the image to be created through LF will be pointing to the right version of drools apps in the config files.   This will work unless for some reason we mess up the drools-application image and we need to bump it by an additonal patch version.   Thoughts.   I'm doing a +2, case you agree with the above Pam, and want to go ahead and release.

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2