More sonar cleanup and line consolidation
Missing javadoc and some variable naming to remove consecutive
capitlization of letters.

Started doing some combining of lines at the risk of making the
code even more unreadable than it already is.

Started addressing some nested if's.

I see opportunity for a lot of common code to be combined or
put into a base controller or utility method.

Issue-ID: POLICY-2133
Change-Id: I6d4d9618bea685cf1d02899e3a9d4a43919a2a96
Signed-off-by: Pamela Dragosh <pdragosh@research.att.com>